@@ -1,23 +1,19 @@
-import logo from './logo.svg';
 import './App.css';
+import { Layout } from './components/Layout'
+import { Home } from './pages/Home'
+import { NextUIProvider } from '@nextui-org/react'
+import { Container } from '@nextui-org/react'
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <NextUIProvider>
+        <Layout>
+          <Container>
+            <Home />
+          </Container>
+        </Layout>
+      </NextUIProvider>
     </div>
   );
 }
